+++ src/test/java/se/bjurr/prnfs/admin/utils/PullRequestRefBuilder.java
@@ -1,14 +1,27 @@
 package se.bjurr.prnfs.admin.utils;
 
+import static com.google.common.collect.Sets.newHashSet;
+import static org.mockito.Mockito.when;
+
+import java.util.Set;
+
+import org.mockito.Matchers;
+
+import se.bjurr.prnfs.listener.PrnfsRenderer;
+import se.bjurr.prnfs.listener.PrnfsRenderer.REPO_PROTOCOL;
+
 import com.atlassian.stash.project.Project;
 import com.atlassian.stash.project.ProjectType;
 import com.atlassian.stash.project.ProjectVisitor;
 import com.atlassian.stash.pull.PullRequestRef;
 import com.atlassian.stash.repository.Repository;
+import com.atlassian.stash.repository.RepositoryCloneLinksRequest;
+import com.atlassian.stash.repository.RepositoryService;
+import com.atlassian.stash.util.NamedLink;
 
 public class PullRequestRefBuilder implements PullRequestRef {
  public static PullRequestRefBuilder pullRequestRefBuilder() {
-  return new PullRequestRefBuilder();
+  return new PullRequestRefBuilder(null);
  }
 
  private String hash;
@@ -19,8 +32,21 @@ public class PullRequestRefBuilder implements PullRequestRef {
  private String repositoryName;
 
  private String slug;
+ private PullRequestEventBuilder pullRequestEventBuilder;
 
  private PullRequestRefBuilder() {
+ }
+
+ private PullRequestRefBuilder(PullRequestEventBuilder pullRequestEventBuilder) {
+  this.pullRequestEventBuilder = pullRequestEventBuilder;
+ }
+
+ public static PullRequestRefBuilder pullRequestRefBuilder(PullRequestEventBuilder pullRequestEventBuilder) {
+  return new PullRequestRefBuilder(pullRequestEventBuilder);
+ }
+
+ public PullRequestEventBuilder build() {
+  return pullRequestEventBuilder;
  }
 
  @Override
@@ -38,8 +64,8 @@ public class PullRequestRefBuilder implements PullRequestRef {
   return hash;
  }
 
+ // Should not use @Override, since its only available from 3.7.0
  public String getLatestCommit() {
-  // Only in 3.7.0
   return hash;
  }
 
@@ -85,14 +111,16 @@ public class PullRequestRefBuilder implements PullRequestRef {
      public Integer getId() {
       return projectId;
      }
-
+     
+     // Should not use @Override, since its only available until 3.12.0
      public boolean getIsPersonal() {
-      // In 2.12.0
       return false;
      }
 
      @Override
      public String getKey() {
+      // Avoid save action from auto remove unused method
+      getIsPersonal();
       return projectKey;
      }
 
@@ -183,5 +211,26 @@ public class PullRequestRefBuilder implements PullRequestRef {
  public PullRequestRefBuilder withRepositorySlug(String repositorySlug) {
   this.slug = repositorySlug;
   return this;
+ }
+
+ public PullRequestRefBuilder withCloneUrl(PrnfsRenderer.REPO_PROTOCOL protocol, String url) {
+  mockCloneUrl(protocol, pullRequestEventBuilder.getPrnfsTestBuilder().getRepositoryService(), url);
+  return this;
+ }
+
+ private void mockCloneUrl(REPO_PROTOCOL protocol, RepositoryService repositoryService, final String url) {
+  NamedLink e = new NamedLink() {
+   @Override
+   public String getName() {
+    return null;
+   }
+
+   @Override
+   public String getHref() {
+    return url;
+   }
+  };
+  Set<NamedLink> value = newHashSet(e);
+  when(repositoryService.getCloneLinks(Matchers.any(RepositoryCloneLinksRequest.class))).thenReturn(value);
  }
 }